requirements.txt
annotated-types==0.5.0
anyio==3.7.1
exceptiongroup==1.1.2
fastapi==0.100.1
h11==0.14.0
httptools==0.6.0
idna==3.4
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
watchfiles==0.19.0
websockets==11.0.3
sqlalchemy==2.0.19
pydantic_settings==2.0.2
psycopg2-binary==2.9.6
python-socketio==5.8.0
redis==5.0.0